Refactor block bindings utils
